fix(studio): quota integer out of range + proxy telechargement + bucket dedie + preview_url
- Trigger quota : 2*1024^3 en litteraux int4 DEBORDAIT (integer out of range)
  -> tout upload echouait. Cast 2::bigint — re-applique et prouve en prod.
- Route GET /api/studio/templates/:id/download : verifyJWT + OWNERSHIP (403
  autrui) + rate-limit ; lit GCS ou bucket Supabase prive via service_role et
  streame -> plus jamais de fetch direct navigateur->GCS (CORS).
- Migration 20260806100000 : preview_url (vignettes) + bucket Supabase DEDIE
  prive studio-templates avec 3 policies owner par prefixe {userId}/ — fini le
  fallback dans communication-files (bucket messagerie). Appliquee prod.

+++ supabase/migrations/20260805200000_provider_uploaded_templates.sql
@@ -52,7 +52,8 @@ SET search_path = public
 AS $$
 DECLARE
   v_total bigint;
-  v_quota constant bigint := 2 * 1024 * 1024 * 1024; -- 2 Go
+  -- ⚠️ 2*1024*1024*1024 en littéraux int4 DÉBORDE (integer out of range) → cast bigint AVANT.
+  v_quota constant bigint := 2::bigint * 1024 * 1024 * 1024; -- 2 Go
 BEGIN
   SELECT COALESCE(SUM(size_bytes), 0) INTO v_total
   FROM public.provider_uploaded_templates
